python-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-docx==1.1.2
reportlab==4.2.2
python-pptx==1.0.2
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
pypdf==6.9.2
pymupdf==1.27.2
python-multipart==0.0.20
